pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "cloudinary>=1.44.1",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "django-unfold>=0.66.0",
    "django-cloudinary-storage>=0.3.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "flower>=2.0.1",
    "gunicorn>=25.3.0",
    "psycopg[binary]>=3.3.3",
    "sentry-sdk>=2.40.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"config/settings/*.py" = ["F403", "F405"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disable_error_code = ["import-untyped", "var-annotated", "attr-defined"]
